unit tests for new factory classes added

+++ randomizer-core/src/main/java/de/alpharogroup/random/number/RandomPrimitivesFactory.java
@@ -1,3 +1,27 @@
+/**
+ * The MIT License
+ *
+ * Copyright (C) 2015 Asterios Raptis
+ *
+ * Permission is hereby granted, free of charge, to any person obtaining
+ * a copy of this software and associated documentation files (the
+ * "Software"), to deal in the Software without restriction, including
+ * without limitation the rights to use, copy, modify, merge, publish,
+ * distribute, sublicense, and/or sell copies of the Software, and to
+ * permit persons to whom the Software is furnished to do so, subject to
+ * the following conditions:
+ *
+ * The above copyright notice and this permission notice shall be
+ * included in all copies or substantial portions of the Software.
+ *
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND,
+ * EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF
+ * MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND
+ * NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT HOLDERS BE
+ * LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION
+ * OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION
+ * WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
+ */
 package de.alpharogroup.random.number;
 
 import java.security.SecureRandom;
@@ -11,7 +35,7 @@ public final class RandomPrimitivesFactory
 {
 	/**
 	 * Returns a random boolean.
-	 * 
+	 *
 	 * @param secureRandom
 	 *            the secure random for boolean generation
 	 *
@@ -25,7 +49,7 @@ public final class RandomPrimitivesFactory
 
 	/**
 	 * The Method randomByte() selects a random byte.
-	 * 
+	 *
 	 * @param secureRandom
 	 *            the secure random for byte generation
 	 *
@@ -295,7 +319,7 @@ public final class RandomPrimitivesFactory
 	 *            the secure random for int generation
 	 * @return an int not greater then the range.
 	 */
-	public static long randomInt(final int range, SecureRandom secureRandom)
+	public static int randomInt(final int range, SecureRandom secureRandom)
 	{
 		return randomInt(range, RandomAlgorithm.SECURE_RANDOM, secureRandom);
 	}
@@ -414,7 +438,7 @@ public final class RandomPrimitivesFactory
 
 	/**
 	 * Returns a random short
-	 * 
+	 *
 	 * @param secureRandom
 	 *            the secure random for short generation
 	 *
